chore: rename client worker, worker and composable modules
