+++ app/utils/pkg-size/analyze-cause-client-worker.ts
@@ -1,4 +1,4 @@
-import Worker from '~/utils/pkg-size/analize-cause-worker?worker'
+import Worker from '~/utils/pkg-size/analyze-cause-worker?worker'
 
 export const worker = new Worker({
   name: 'NpmxPkgSizeAnalyzeCauseWorker',
